feat: implement fsd
