+++ src/app/types/globals.d.ts
