--- src/components/Counter/Counter.module.css
@@ -1,5 +1,0 @@
-.wrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
