--- src/components/Counter/Counter.tsx
@@ -1,23 +1,0 @@
-import { useState } from "react";
-import styles from "./Counter.module.css";
-
-const Counter = () => {
-  const [count, setCount] = useState(0);
-
-  const increment = () => {
-    setCount((oldCount) => oldCount + 1);
-  };
-
-  const decrement = () => {
-    setCount((oldCount) => oldCount - 1);
-  };
-
-  return (
-    <div className={styles.wrapper}>
-      <div>{count}</div>
-      <button onClick={increment}>Increment</button>
-      <button onClick={decrement}>Decrement</button>
-    </div>
-  );
-};
-export default Counter;
