+++ src/index.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
-import { App } from "./App";
-import { ThemeProvider } from "./providers/ThemeProvider";
+import { App } from "./app/App";
+import { ThemeProvider } from "./app/providers/ThemeProvider";
 const container = document.getElementById("root");
 const root = createRoot(container!);
