+++ src/pages/AboutPage/index.ts
@@ -1,0 +1,1 @@
+export { AboutPageLazy as AboutPage } from "./ui/AboutPage.async";
