+++ src/shared/contexts/theme/theme-context.ts
@@ -1,0 +1,18 @@
+import { createContext } from "react";
+
+export enum Theme {
+  LIGHT = "light",
+  DARK = "dark",
+}
+
+export interface ThemeContextProps {
+  theme: Theme;
+  setTheme: (newTheme: Theme) => void;
+}
+
+export const ThemeContext = createContext<ThemeContextProps>({
+  theme: Theme.LIGHT,
+  setTheme: () => {},
+});
+
+export const LOCAL_STORAGE_THEME_KEY = "theme";
