+++ tsconfig.json
@@ -1,6 +1,9 @@
 {
   "compilerOptions": {
     "baseUrl": ".",
+    "paths": {
+      "*": ["./src/*"]
+    },
     "outDir": "./build/",
     "noImplicitAny": true,
     "strict": true,
